built the skeletal structure for things to come
The `nix` directory holds the necessary nix code to deterministically
build all assets in this project. To that end, `nix/default.nix` exports
the `nixpkgs` attribute-set used by this project, customized with the
overlays defined in `nix/overlays`.

The `shell.nix` file defines the developer environment that can be
accessed by running `nix-shell` from the root of this repository. To be
extra careful, you may run `nix-shell --pure` to ensure an absolutely
pure development environment.