use npmlock2nix instead of node2nix; remove ununsed nix overlays (#39)
* use npmlock2nix instead of node2nix; remove ununsed nix overlays

* nix: npm2nix to a WIP lockfiles v2 branch

* move evm contracts into its own directory

* nix/portal: build the project with nodejs-16 instead of nodejs-17.

* nix/portal/contracts: set HOME for truffle build

Truffle is trying to create XDG_CONFIG (resolved at $HOME/.config on
linux) at startup. It doesn't seem to write anything in the directory.

Setting HOME to the nix build worker sandbox temporary root to prevent
truffle from trying to create a directory in /homeless-shelter/.

* nix/portal/contracts: use nixpkgs-provided solc

By default, Truffle is trying to fetch the solc compiler from the
internet using some heuristics to determine what kind of binary the
host system is looking for.

This is considered as a bad practice in terms of bootstrapping and
reproducible builds. Who built the compiler you're using? Is it really
built from the source it pretends to be built from? There's no way to
tell.

Nix is preventing this kind of bad practices by preventing build
sandboxes to access to the internet.

Instead of remotely fetching the solidity compiler, we inject
the Nixpkgs source-built one. It gives us the certainty that the
solidity compiler we're using has been built from the source code it
pretends to be coming from :)

-----------------------------------------------

I hit a weird bug leading me to relax the solidity pragma. I'm not
familiar with solc enough to pinpoint the exact issue though :/

We're using solc 0.8.2, the pragma requires something >0.5.8. Yet, the
version bound check seems to fail with:

```
CompileError: ParserError: Source file requires different compiler version (current compiler is 0.8.2+commit.661d1103.Linux.g++) - note that nightly builds are considered to be strictly less than the released version
 --> project:/contracts/swap.sol:2:1:
  |
2 | pragma solidity ^0.5.8;
  | ^^^^^^^^^^^^^^^^^^^^^^^

Error: Truffle is currently using solc native, but one or more of your contracts specify "pragma solidity ^0.5.8".
```

I feel like the `+commit.661d1103.Linux.g++` is the culprit here.
Looking at the Nixpkgs
derivation (https://github.com/NixOS/nixpkgs/blob/master/pkgs/development/compilers/solc/default.nix),
it's not clear to me where this "commit.xxx" bit is coming from. We'd
probably need to dig into the solc source code to some extent to
understand what's happening here.

I don't think we want to jump that rabbit hole just yet, I instead
relaxed the version bound.

Note: we could relax this through a source substitution in the
nix-build, leaving the bound in the repo file intact.

* enable smart contract testing

* remove dead code

* [nix] Fix the truffle situation

There was some tension between the interactive (ie. non-nix) truffle
usage and the nix-build truffle one. When used interactively, we want
truffle to figure out the solc installation, at least for now. When
ran inside of a Nix build, we want truffle to use the nixpkgs-provided
solc binary.

To fix this unfortunate conflict, we extract the truffle configuration
in a separate JSON file. We leave the interactive configuration in the
repository. IE. the devs/non-nix people will be able to immediately
use the repository.

As for the nix build, we patch the truffle configuration JSON file to
force truffle to use the nixpkgs-provided solc.

Note: the original configuration will remain untouched, we're just
patching the source present in the nix builder sandbox.

------------------

We also patch the shebangs of the ganache scripts with the appropriate
nodejs-6 one.

* [nix] patch the solidity pragmas

The nixpkgs-provided solc seems to have some issue to check its own
version against a pragma-provided one. See the
https://github.com/FabricLabs/mono/pull/39/commits/b657a59344a01ec9734f46720a19c38c041bd1c0
commit message for more details

We restore the original contract pragma and relax it during the
prePatch phase of the nix build instead.

* [nix] Setup github actions

* [nix] bump nixpkgs

The MacOS solc problem has been "fixed" upstream:
- https://github.com/NixOS/nixpkgs/commit/fb60c5e7b3273d165ff682be81265b3ef301dee2

Note: it's a bindist, but hey, it's still better than a broken build!

* [nix] Setup dev shells for the contract and portal projects

The diff look scary, but it's mostly stuff being moved around.

We move the projects nix definitions in their own /js directories. We
then import them in the portal overlay.

We then add a nix-shell setup for both the portal and the contracts
compilation project.

For the contracts project, we inject the nixpkgs-provided solc binary
through a env variable pointing to a patched truffle configuration.
This approach leaves the checked-in truffle configuration untouched.

* [nix] add portaldefi namespace to nixpkgs overlay

We add a portaldefi namespace to make sure our local packages won't
collide with Nixpkgs in the future.

* fix vendor hash for terraform-provider-equinix

* minor fixes and nit-picky refactoring

* regenerate npm dependencies and locks

Co-authored-by: Félix Baylac-Jacqué <felix@alternativebit.fr>