Continuous Integration and Deployment (#42)
* initial commit

* [nix] add portal unit test + ci target

* [portal] Npm-Build the project using --production

This flag will prevent npm from packing the dev dependencies to the
node_modules directory.

* [nix] Write a NixOS VM Test for the portal

* [ci] Run tests and vm tests on the CI

* toremove: temporarily enabling deploy workflow to test it

* Unit and Integration testing (#43)

* [nix] Fix unit-test target

* [CI] Move unit-tests run to build job.

Each Github job requires a full nix new install and will empty the
store. We'll have to re-populate the store from the cache, which tend
to be a pretty expensive operation (~1.5 min for each job ATM).

Let's try to keep the number of jobs as a bare minimum. At least until
we get a runner having a persistant store.

We disable the VM test for the macos job: NixOS VM tests can't run on
MacOS for now.

* [contracts] Remove solidity pragma

* [nix] update npmlock2nix pin

Reflecting the npm-extra-args branch rebase to include the latest
bugfixes.

See the following commit messages for more informations:
- https://github.com/NinjaTrappeur/npmlock2nix/commit/330f8d574ab58cb6413141aa62bab3330871a365
- https://github.com/NinjaTrappeur/npmlock2nix/commit/5c86607681a84702d94f213e2fec900890ca0cd7

* [portal] use --omit=dev instead of --production

Npm is issuing a deprecation warning asking as to replace --production
with --omit=dev. Let's comply with it.

* [nix] centralize the nodejs version

This makes it easier to use the same version everywhere.

* [nix] use the same shell style as in the workshop

The previous approach is also valid. I'm just trying to align things a
bit.

* [nix] unmock VM test

We were mocking the portal server in the VM test to go around the
broken portal.

Now the portal has been fixed, let's unmock that.

The test is failing for now: the /api/v1/fees/recommended endpoint
returns a 500. I assume it fails because of the lack of internet
connectivity.

I assume we'll have to mock the data for this test as well if we want
it to succeed.

* [nix] Reflecting portal module rename on nginx

* [nix] remove top-level portal config

* Merge branch 'feature/eth-swaps' into anand/ci-cd

* split tests into unit and integration tests

* [nix] Fix nodejs circular dependency

cd60fbb centralized the nodejs version to the portaldefi overlay.
However, this overlay needs to first evaluate the portal and contracts
projects before getting constructed. Meaning we can't use the overlay
itself from these two projects.

Propagating the nodejs version through the contracts/portal input attrset.

* [nix][portal] Only run unit tests during nix build

* [ci] Run the portal integration tests on CI

These tests need full network access. We can't run them inside of a
nix-build. Using nix-shell to populate the runtime env of the test.

* [portal] Add an alive endpoint to the v1 API.

We use this endpoint to check whether or not the portal service is
alive in the VM tests. Provided that the VM tests don't have any
network access, we need a dummy endpoint not relying on any network
access to check the service health.

* [nix] Set HOME to build dir in portal unit test

For some reason, it seems like mocha is behaving differently on MacOS
and tries to create a tmp directory in the user $HOME.

* [CD] Initial CD setup

We merge the PR and Deploy workflows to easily create a dependency
between the CD and the CI.

We introduce a basic CD system in charge of deploying the NixOS
closure to the playnet server. We tried to stay as simple as possible:
we just copy the machine closure to the remote server before
activating it.

We deploy the master branch to playnet for now.

Co-authored-by: Félix Baylac-Jacqué <felix@alternativebit.fr>
Co-authored-by: zimbatm <zimbatm@zimbatm.com>